[RUCIO-2777] DB modules were wrongfully imported in common.utils
